test_diplom/requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
docopt==0.6.2
flake8==7.2.0
greenlet==3.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.2
pipreqs==0.4.13
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarg==0.1.10
